[Refactor] : 폴더 구조 정리 공용 컴포넌트는 all 폴더로 이동
