+++ src/components/all/HistoryComponentUpperSection.tsx
