+++ src/components/all/ResultCard.tsx
