+++ src/components/payment/after/PaymentAfterConfirm.tsx
@@ -1,7 +1,7 @@
 // 백엔드로부터 예약 확정 정보를 받아오는 것이면 그냥 클라이언트 컴포넌트가 되어도 상관 없음
+import HistoryComponentUpperSection from '@components/all/HistoryComponentUpperSection';
 import FortyEightTigSVG from '@public/svg/fortyEightTig.svg';
 // 하지만 사용자가 새로고침하면 전역 상태라고 하더라도 사라지기 때문에 백엔드로부터 받아오는 것이 더 맞지 않나싶음. 아니면 zustand persists
-import HistoryComponentUpperSection from '@components/reservation-list/all/HistoryComponentUpperSection';
 
 interface paymentAfterConfirmProp {
   reservationId: string;
