+++ src/components/reservation/GameCountCard.tsx
@@ -1,5 +1,5 @@
-import ChooseCard from '@components/search/ChooseCard';
-import InfoCard from '@components/search/InfoCard';
+import InfoCard from '@components/all/InfoCard';
+import ChooseCard from '@components/reservation/ChooseCard';
 import { ChooseCardProps } from 'types/search/SearchTypes';
 
 const chooseLists: ChooseCardProps[] = [
@@ -12,9 +12,7 @@ export default function GameCountCard() {
   return (
     <section className="w-full flex flex-col p-5 mt-5 border-b border-grey2">
       <InfoCard number={4} content="원하는 게임 수를 선택해주세요." />
-      <ChooseCard
-        title="게임"
-      />
+      <ChooseCard title="게임" />
     </section>
   );
 }
