+++ src/components/reservation/ResReservationCard.tsx
@@ -1,9 +1,9 @@
 'use client';
-import InfoCard from '@components/search/InfoCard';
 import { useSearchParams } from 'next/navigation';
 import DateSVG from '@public/svg/date.svg';
 import ResTimeSVG from '@public/svg/resTime.svg';
 import PeopleSVG from '@public/svg/people.svg';
+import InfoCard from '@components/all/InfoCard';
 
 interface ResReservationCardProps {
   date: string;
